requirements.txt
# Core Framework
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic==2.9.2
pydantic-settings==2.5.2
email-validator==2.2.0

# Database
motor==3.6.0
pymongo==4.9.1

# PDF Processing
PyMuPDF==1.24.10
pdfminer.six==20231228
pypdf==5.1.0

# Image Processing
Pillow==11.0.0
pytesseract==0.3.13

# Machine Learning / Embeddings
numpy==2.1.2
scikit-learn==1.5.2

# Google Cloud (for Vertex AI)
google-cloud-aiplatform==1.70.0
google-cloud-storage==2.18.2
google-generativeai==0.8.3

# Authentication
python-jose[cryptography]==3.3.0
passlib==1.7.4
bcrypt==4.0.1
python-multipart==0.0.12

# HTTP Client
httpx==0.27.2
aiohttp==3.10.10

# Utilities
python-dotenv==1.0.1
python-magic==0.4.27

# Logging
structlog==24.4.0

# Security
cryptography==43.0.1

# Testing
pytest==8.3.3
pytest-asyncio==0.24.0
pytest-cov==5.0.0
pytest-mock==3.14.0

# Development
black==24.10.0
flake8==7.1.1
mypy==1.13.0
isort==5.13.2
pre-commit==4.0.1

# Type Stubs
types-requests==2.32.0.20241016
